+++ apps/protoform/src/app/credit-cards/credit-limit/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ButtonGroup, Form, FormGroup, Input, InputGroup } from '@westpac/ui';
+import { ButtonGroup, ButtonGroupButton, Field, Form, FormGroup, Input, InputGroup } from '@westpac/ui';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useCallback, useEffect } from 'react';
 import { Controller, useForm } from 'react-hook-form';
@@ -75,22 +75,23 @@ Choose your credit card limit or have one allocated for you."
             control={control}
             name="creditLimitType"
             rules={{ required: defaultError }}
-            render={({ field: { onChange, onBlur, value } }) => (
-              <ButtonGroup
+            render={({ field: { onChange, value } }) => (
+              <Field
                 label="What credit limit would you like?"
                 hintMessage="You can chose a limit between $500 and $4,000 for Westpac Lite Visa Card."
-                size="large"
-                onChange={onChange} // send value to hook form
-                onBlur={onBlur} // notify when input is touched/blur
-                value={value}
-                block={{ initial: true, md: false }}
                 errorMessage={errors.creditLimitType?.message}
-                defaultValue={data.creditLimitType}
-                buttons={[
-                  { value: 'Allocate', label: 'Allocate for me' },
-                  { value: 'Specify', label: 'Specify' },
-                ]}
-              />
+              >
+                <ButtonGroup
+                  size="large"
+                  onSelectionChange={onChange} // send value to hook form
+                  selectedKeys={value}
+                  block={{ initial: true, md: false }}
+                  defaultSelectedKeys={data.creditLimitType}
+                >
+                  <ButtonGroupButton id="Allocate">Allocate for me</ButtonGroupButton>
+                  <ButtonGroupButton id="Specify">Specify</ButtonGroupButton>
+                </ButtonGroup>
+              </Field>
             )}
           />
         </FormGroup>