+++ packages/ui/.storybook/main.ts
@@ -4,7 +4,6 @@ import { dirname, join } from 'node:path';
 import { mergeConfig } from 'vite';
 
 import type { StorybookConfig } from '@storybook/react-vite';
-import { get } from 'node:http';
 
 const require = createRequire(import.meta.url);
 
@@ -28,9 +27,6 @@ const config: StorybookConfig = {
 
   viteFinal(config) {
     return mergeConfig(config, {
-      optimizeDeps: {
-        exclude: ['@duetds/date-picker/dist/loader'],
-      },
       define: {
         'process.env': {},
       },